requirements/base.txt
pytz==2018.7  # https://github.com/stub42/pytz
python-slugify==2.0.0  # https://github.com/un33k/python-slugify
Pillow==5.3.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
redis>=2.10.6, < 3  # pyup: < 3 # https://github.com/antirez/redis
selenium==3.141.0  # https://semaphoreci.com/community/tutorials/setting-up-a-bdd-stack-on-a-django-application
zappa==0.47.1  # https://github.com/Miserlou/Zappa#installation-and-configuration
tablib==0.12.1  # http://docs.python-tablib.org/en/master/tutorial/
pyquery==1.4.0  # https://pythonhosted.org/pyquery/manipulating.html

# Django
# ------------------------------------------------------------------------------
django==2.0.9  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.38.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis
django-s3-storage==0.12.4
django-admin-sortable2==0.6.21
django-polymorphic==2.0.3
django-autocomplete-light==3.3.2
django-extensions==2.1.4  # https://github.com/django-extensions/django-extensions


# Django REST Framework
djangorestframework==3.9.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
